Create separate component for NylasLogin
